Refactoring TabNavigation
