+++ src/navigation/AppNavigator.tsx
@@ -1,28 +1,8 @@
 import React from 'react';
 import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
 import { NavigationContainer } from '@react-navigation/native';
-import { createStackNavigator, StackNavigationProp } from '@react-navigation/stack';
-import HomeScreen from '../screens/HomeScreen';
 import CreateTaskScreen from '../screens/CreateTaskScreen';
-import TaskDetailScreen from '../screens/TaskDetailScreen';
-import { Task } from '../types/Task';
-
-export type RootStackParamList = {
-  Tareas: undefined;
-  Detalles: { task: Task };
-  CrearTarea: undefined;
-};
-
-export type TaskDetailNavigationProp = StackNavigationProp<RootStackParamList, 'Detalles'>;
-
-const Stack = createStackNavigator();
-
-const TasksStack = () => (
-  <Stack.Navigator>
-    <Stack.Screen name="Tareas" component={HomeScreen} />
-    <Stack.Screen name="Detalles" component={TaskDetailScreen} />
-  </Stack.Navigator>
-);
+import TasksStack from './TasksStack';
 
 const Tab = createBottomTabNavigator();
