+++ src/types/Task.ts
@@ -1,3 +1,5 @@
+import { StackNavigationProp } from "@react-navigation/stack";
+
 export interface Task {
   id: number;
   title: string;
@@ -6,7 +8,18 @@ export interface Task {
   completed: boolean;
 }
 
+//StackNavigation
 export interface TaskDetailProps {
   route: { params: { task: Task } };
   navigation: { goBack: () => void };
 }
+
+//Nuestro RootNavigation BottomTabs
+export type RootStackParamList = {
+  Tareas: undefined;
+  Detalles: { task: Task };
+  CrearTarea: undefined;
+};
+
+//Funciona para errores de tipo del Stack
+export type TaskDetailNavigationProp = StackNavigationProp<RootStackParamList, 'Detalles'>;
